Merge develop into master

+++ app/Models/Actionlog.php
@@ -295,13 +295,13 @@ class Actionlog extends SnipeModel
         $now = Carbon::now();
         $last_audit_date = $this->created_at; // this is the action log's created at, not the asset itself
         $next_audit = $last_audit_date->addMonth($monthInterval); // this actually *modifies* the $last_audit_date
-        $next_audit_days = round($now->diffInDays($next_audit, true));
+        $next_audit_days = (int) round($now->diffInDays($next_audit, true));
         $override_default_next = $next_audit;
 
         // Override the default setting for interval if the asset has its own next audit date
         if (($asset) && ($asset->next_audit_date)) {
             $override_default_next = Carbon::parse($asset->next_audit_date);
-            $next_audit_days = round($override_default_next->diffInDays($now, true));
+            $next_audit_days = (int) round($override_default_next->diffInDays($now, true));
         }
 
         // Show as negative number if the next audit date is before the audit date we're looking at
